Added testDispose() and removed extra InventoryDisplayTest2.java
